requirements.txt
# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
audioread>=3.0.0

# Numerical Computing
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0

# Machine Learning
scikit-learn>=1.3.0
umap-learn>=0.5.0
hdbscan>=0.8.33

# Data Storage
pyarrow>=14.0.0

# Visualization (for verification)
matplotlib>=3.7.0

# Web UI
flask>=3.0.0
flask-cors>=4.0.0

# Testing
pytest>=7.4.0
